refac: rename getDhatupatha to getDhatupathaLocal
Why?
Going to add `getDhatupatha` which is going to run on client-side.
