+++ src/utils/search/sortData.ts
@@ -1,4 +1,4 @@
-import { DhatuDetails } from "../getDhatupatha";
+import { DhatuDetails } from "../getDhatupathaLocal";
 import { sortOptions } from "../viewInputsData";
 import { SortQuery } from "./getSortQuery";
